chore: Refactor humeur command to validate and handle image URLs
This commit refactors the `humeur` command in the `humeur.js` file to validate and handle image URLs as input. Previously, the command only accepted predefined mood values (happy, neutral, sad), but now it also allows users to provide a URL of an image. The change includes validation checks to ensure that the URL is valid and returns an image. If the URL is invalid, an error message is displayed to the user.